test: split orchestration tests and colocate test files with source units
